+++ src/scripts/admin.js
@@ -2,6 +2,7 @@ import '../styles/admin.scss';
 import '../styles/public.scss';
 import './admin/hazard-type-select';
 import './admin/hide-block';
+import './admin/region-group';
 
 /**
  * All the scripting for the admin area goes here.